1 commit, changed points class to own file to make code nicer to read and aslo to work out how to do this. Python testing after all
